Consolidate content validation types under contentvalidation package (#48272)
