+++ sdk/storage/azure-storage-common/src/main/java/com/azure/storage/common/implementation/contentvalidation/StorageCrc64Calculator.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.storage.common.implementation;
+package com.azure.storage.common.implementation.contentvalidation;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
