+++ sdk/storage/azure-storage-common/src/main/java/com/azure/storage/common/implementation/contentvalidation/package-info.java
@@ -4,4 +4,4 @@
 /**
  * Package containing classes for structured message encoding and decoding.
  */
-package com.azure.storage.common.implementation.structuredmessage;
+package com.azure.storage.common.implementation.contentvalidation;
